+++ test/send.js
@@ -147,14 +147,21 @@ describe('send(file).pipe(res)', function () {
   it('should set Content-Type via mime map', function (done) {
     request(app)
       .get('/name.txt')
-      .expect('Content-Type', 'text/plain; charset=UTF-8')
+      .expect('Content-Type', 'text/plain; charset=utf-8')
       .expect(200, function (err) {
         if (err) return done(err)
         request(app)
           .get('/tobi.html')
-          .expect('Content-Type', 'text/html; charset=UTF-8')
+          .expect('Content-Type', 'text/html; charset=utf-8')
           .expect(200, done)
       })
+  })
+
+  it('should default Content-Type to octet-stream', function (done) {
+    request(app)
+      .get('/no_ext')
+      .expect('Content-Type', 'application/octet-stream')
+      .expect(200, done)
   })
 
   it('should 404 if file disappears after stat, before open', function (done) {
@@ -1277,40 +1284,6 @@ describe('send(file, options)', function () {
           .get('/do..ts.txt')
           .expect(200, '...', done)
       })
-    })
-  })
-})
-
-describe('send.mime', function () {
-  it('should be exposed', function () {
-    assert.ok(send.mime)
-  })
-
-  describe('.default_type', function () {
-    before(function () {
-      this.default_type = send.mime.default_type
-    })
-
-    afterEach(function () {
-      send.mime.default_type = this.default_type
-    })
-
-    it('should change the default type', function (done) {
-      send.mime.default_type = 'text/plain'
-
-      request(createServer({ root: fixtures }))
-        .get('/no_ext')
-        .expect('Content-Type', 'text/plain; charset=UTF-8')
-        .expect(200, done)
-    })
-
-    it('should not add Content-Type for undefined default', function (done) {
-      send.mime.default_type = undefined
-
-      request(createServer({ root: fixtures }))
-        .get('/no_ext')
-        .expect(shouldNotHaveHeader('Content-Type'))
-        .expect(200, done)
     })
   })
 })